+++ src/components/ids-module-nav/demos/index.ts
@@ -1,0 +1,7 @@
+import '../ids-module-nav';
+import '../ids-module-nav-bar';
+import '../ids-module-nav-button';
+import '../ids-module-nav-content';
+import '../ids-module-nav-item';
+import '../ids-module-nav-settings';
+import '../ids-module-nav-switcher';